🤡 remove function from commands and add in utils folder
